requirements.txt
# ── Core LLM (FREE via Groq) ─────────────────────────────────
groq==0.11.0

# ── Web framework ────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
aiofiles==24.1.0

# ── Data validation ──────────────────────────────────────────
pydantic==2.9.2

# ── Environment & utils ──────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.2

# ── GitHub integration (GitHub Apps are FREE) ────────────────
pygithub==2.4.0

# ── Local database (SQLite, no server needed) ────────────────
# Built into Python — no install needed

httpx==0.27.2